dags/energy_data_pipeline_starter.py
"""
Student Data Engineer Challenge - Energy Data Pipeline
Terra One Coding Challenge

This is a STARTER template for your Airflow DAG.
Complete the TODOs to implement the full pipeline.

Author: [Your Name]
Date: [Today's Date]
"""

from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.operators.bash import BashOperator
import pandas as pd
import numpy as np
import logging
import os
from pathlib import Path

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Default DAG arguments
default_args = {
    'owner': 'student-data-engineer',
    'depends_on_past': False,
    'start_date': datetime(2024, 1, 1),
    'email_on_failure': False,
    'email_on_retry': False,
    'retries': 1,
    'retry_delay': timedelta(minutes=5),
}

# Initialize DAG
dag = DAG(
    'energy_data_pipeline',
    default_args=default_args,
    description='Energy market data processing pipeline for Terra One challenge',
    schedule_interval=None,  # Manual trigger for challenge
    catchup=False,
    tags=['student-challenge', 'energy-data', 'terra-one'],
)

# Data paths - modify these if needed
RAW_DATA_PATH = '/opt/airflow/dags/energy_data.csv'
PROCESSED_DATA_DIR = '/opt/airflow/processed_data'
OUTPUT_DIR = '/opt/airflow/output'

def validate_data_quality(**context):
    """
    Task 1: Data Validation & Quality Check
    
    TODO: Implement data quality validation and generate a report
    """
    logger.info("Starting data quality validation...")
    
    # TODO: Load data
    # df = pd.read_csv(RAW_DATA_PATH)
    
    # TODO: Implement your validation logic here
    
    # TODO: Generate and save quality report
    quality_report = {
        'total_records': 0,  # TODO: Replace with actual count
        'missing_values': {},  # TODO: Count per column
        'data_issues': [],  # TODO: List issues found
        'validation_passed': True  # TODO: Set based on your criteria
    }
    
    logger.info(f"Data quality report: {quality_report}")
    
    # Return for downstream tasks to use
    return quality_report

def clean_and_standardize_data(**context):
    """
    Task 2: Data Cleaning & Standardization
    
    TODO: Implement data cleaning logic
    """
    logger.info("Starting data cleaning and standardization...")
    
    # TODO: Load raw data
    # df = pd.read_csv(RAW_DATA_PATH)
    
    # TODO: Implement your cleaning steps here
    
    # TODO: Save cleaned data
    # cleaned_df.to_csv(f"{OUTPUT_DIR}/cleaned_data.csv", index=False)
    
    logger.info("Data cleaning completed")
    return "Data cleaned successfully"

def partition_and_store_data(**context):
    """
    Task 3: Data Partitioning & Storage
    
    TODO: Partition cleaned data by day and save as Parquet files
    """
    logger.info("Starting data partitioning...")
    
    # TODO: Implement partitioning logic
    
    logger.info("Data partitioning completed")
    return "Data partitioned and stored"

def calculate_hourly_aggregations(**context):
    """
    Task 4: Hourly Aggregations
    
    TODO: Calculate hourly metrics for each day
    """
    logger.info("Calculating hourly aggregations...")
    
    # TODO: Implement aggregation logic
    
    logger.info("Hourly aggregations completed")
    return "Hourly aggregations calculated"

def simulate_market_clearing(**context):
    """
    Task 5: Market Analysis - Market Clearing Simulation
    
    TODO: Implement simplified market clearing
    """
    logger.info("Running market clearing simulation...")
    
    # TODO: Implement market clearing logic
    
    logger.info("Market clearing simulation completed")
    return "Market clearing completed"

def renewable_impact_analysis(**context):
    """
    Task 6: Renewable Energy Impact Simulation
    
    TODO: Simulate renewable energy scenarios
    """
    logger.info("Starting renewable impact analysis...")
    
    # TODO: Implement renewable scenarios
    
    logger.info("Renewable impact analysis completed")
    return "Renewable scenarios analyzed"

def generate_final_summary(**context):
    """
    Task 7: Output Generation
    
    TODO: Create final summary CSV with all metrics
    """
    logger.info("Generating final summary...")
    
    # TODO: Implement summary generation
    
    logger.info("Final summary generated")
    return "Summary report created"

def setup_monitoring(**context):
    """
    Task 8: Pipeline Monitoring
    
    TODO: Implement data quality checks and alerts
    """
    logger.info("Setting up monitoring and alerts...")
    
    # TODO: Implement monitoring logic
    
    logger.info("Monitoring setup completed")
    return "Monitoring configured"

# Create directory structure
create_directories = BashOperator(
    task_id='create_directories',
    bash_command=f'mkdir -p {PROCESSED_DATA_DIR} {OUTPUT_DIR}',
    dag=dag,
)

# Task 1: Data Quality Validation
validate_quality = PythonOperator(
    task_id='validate_data_quality',
    python_callable=validate_data_quality,
    dag=dag,
)

# Task 2: Data Cleaning
clean_data = PythonOperator(
    task_id='clean_and_standardize_data', 
    python_callable=clean_and_standardize_data,
    dag=dag,
)

# Task 3: Data Partitioning
partition_data = PythonOperator(
    task_id='partition_and_store_data',
    python_callable=partition_and_store_data,
    dag=dag,
)

# Task 4: Hourly Aggregations
hourly_aggs = PythonOperator(
    task_id='calculate_hourly_aggregations',
    python_callable=calculate_hourly_aggregations,
    dag=dag,
)

# Task 5: Market Clearing
market_clearing = PythonOperator(
    task_id='simulate_market_clearing',
    python_callable=simulate_market_clearing,
    dag=dag,
)

# Task 6: Renewable Impact
renewable_analysis = PythonOperator(
    task_id='renewable_impact_analysis',
    python_callable=renewable_impact_analysis,
    dag=dag,
)

# Task 7: Final Summary
final_summary = PythonOperator(
    task_id='generate_final_summary',
    python_callable=generate_final_summary,
    dag=dag,
)

# Task 8: Monitoring Setup
monitoring = PythonOperator(
    task_id='setup_monitoring',
    python_callable=setup_monitoring,
    dag=dag,
)

# Define task dependencies
create_directories >> validate_quality >> clean_data >> partition_data
partition_data >> hourly_aggs >> market_clearing >> renewable_analysis
renewable_analysis >> final_summary
final_summary >> monitoring
